FS_MAGIC: a full 48-bit pattern, s_magic unsigned
- 0xBE50006F11E5 does not fit the target's 41-bit signed int, so s_magic
  is `unsigned' and fsutil holds the magic as a raw Word, outside
  to_word()/from_word(); b6fsutil now prints it in hex.
- const.s: the literal pool carried the old octal magic; the new one takes
  its place there instead of allocating a const word.
- kernel/CMakeLists.txt: const.s is #included by besm6.S and was in no
  dependency list, so editing it rebuilt nothing.

+++ cmd/fsutil/create.cpp
@@ -79,7 +79,7 @@ void create_filesystem(Filesystem &fs, const std::string &path, int64_t nblk, in
     fs.image.create(path, nblk);
 
     fs.sb        = SuperBlock{};
-    fs.sb.magic  = int64_t(FS_MAGIC);
+    fs.sb.magic  = FS_MAGIC;
     fs.sb.bsize  = BSIZEW;
     fs.sb.inopb  = INOPB;
     fs.sb.naddr  = NADDR;
